requirements.txt
aiohttp==3.13.2
albumentations==2.0.8
beautifulsoup4==4.14.2
boto3==1.40.67
botocore==1.40.67
click==8.3.0
doclayout_yolo==0.0.4
fast_langdetect==1.0.0
fastapi==0.121.0
fastmcp==2.13.0.2
fasttext_predict==0.9.2.4
flash_attn==2.8.3
ftfy==6.3.1
fuzzywuzzy==0.18.0
gradio==5.49.1
gradio_pdf==0.0.22
huggingface_hub==0.36.0
json_repair==0.52.4
litserve==0.2.16
loguru==0.7.3
magika==1.0.1
markitdown==0.1.3
mineru_vl_utils==0.1.15
minio==7.2.18
mlx_vlm==0.3.5
modelscope==1.31.0
numpy==2.3.4
omegaconf==2.3.0
onnxruntime==1.23.2
openai==2.7.1
opencv_python==4.12.0.88
opencv_python_headless==4.12.0.88
packaging==25.0
pdfminer_six==20250506
pdftext==0.6.3
Pillow==12.0.0
pyclipper==1.3.0.post6
pydantic==2.12.4
pypdf==6.1.3
pypdfium2==4.30.0
pytesseract==0.3.13
python-dotenv==1.2.1
PyYAML==6.0.3
rapid_table==3.0.1
reportlab==4.4.4
Requests==2.32.5
scikit_learn==1.7.2
scipy==1.16.3
Shapely==2.1.2
six==1.17.0
skimage==0.0
starlette==0.50.0
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.0
torch_npu==2.8.0
torchvision==0.24.0
tqdm==4.67.1
transformers==4.57.1
ultralytics==8.3.225
uvicorn==0.38.0
vllm==0.11.0
